fix homepage_url
The URL must start with https:// or http:// otherwise a relative URL is generated, which is not valid.
This is also one of the reasons, why clicking on the `Math Mode` title in `Config > Plugins` does not
open the web page.
The README.md in the plugins repo also has the wrong (realtive) link that resukts in a 404.