fix: Remove multiplexing mode
This plugin used a multiplexing mode to ensure that the inside of
math blocks were styled more like text, and not whatever cm-comment
does.
Unfortunately, the multiplexing mode was interfering with the code
block background colour styling.
The multiplexing mode was never needed anyways, because we do our own
parsing and use that system to add line classes.

The commit removes the multiplexing mode and adds an additional line
class that allows us to preserve the text styling as it was originally
intended.