fix: Prevent top level rows from becoming aligned
fixes #18
I was only able to reproduce this on a macOS computer.
It appears that the lines following a math-input line were being wrapped
up in the `float: left` and aligned themselves to it.
By setting overflow to auto the layout engine is able to correct itself.
I suspect this issue won't exist in future versions of electron, but
here we are.