Native CodeMirror 6 support (#26)
* Initial native CodeMirror 6 support

* Refactoring and performance improvements

Includes:
- Only rerender all math when content that seems to be in a math block
  has changed.
- Improved styling for inline math results.

* reviewability: Revert changes to package-lock.json

* Fix refresh issues, fix left-aligned math

* snake_case

* Fix cursor jumps to the right while editing right-aligned math

* Improve commenting

* Respect Joplin's editor maximum width setting

* Fix CodeMirror 5 regression: Config results shown as "undefined"

This was caused by a refactor that was not fully applied. In
codeMirror5Extension.ts, refresh_widgets checked lineData.isConfig
to determine if a line was a config line. This, however, had been
replaced with a ".type" property.

Because `lineData` had type `any` in `codeMirror5Extension.ts`, this did
not cause a compiler error.

This commit:
1. Gives lineData a type to prevent similar future errors.
2. Replaces the old .isConfig with .type === LineDataType.Config.

Note: The original refactoring was done to separate the different line
data types into separate TypeScript `interface`s and to allow TypeScript
to distinguish between them with type narrowing.

* Fix total variable, fix global: yes blocks preserved after changing notes

This commit:
- Replaces two Object.assigns with {... }s.
   - Object.assign modifies its first argument (rather than creating a
     copy).
   - This caused objects that should have been copied to be modified.
- Removes a default { total: '' } from the default scope, which
  overwrote the total: block_total assignment.