Fix config fetching in the codemirror 5 extension
The getMessage route wasn't initialized yet causing the plugin to fail
on startup. This wasn't an issue previously because I was initiallizing
the getMessage code before registering the content script, but that was
causing a warning in the logs. The previous commit re-ordered the
initialization but didn't fix anything on this end. I didn't notice
until now because the codemirror6 extension doesn't seem to run as
eagerly, so it doesn't trigger the bug.